pyproject.toml
[project]
name = "PatientHub"
version = "0.1.0"
description = "A unified hub to create, simulate, and evaluate methods for patient/client simulation."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "litellm>=1.81.0",
    "instructor>=1.14.4",
    "hydra-core>=1.3.2",
    "pydantic>=2.9",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.6",
    "pandas>=2.3.3",
    "burr[start]>=0.40.2",
    "rich>=14.2.0",
    "typer>=0.12.0",
    "pytest>=9.0.3",
]

[project.optional-dependencies]
dev = [
    "black>=25.9.0",
    "jupyter>=1.1.1",
    "ipykernel>=7.0.1",    
    "chainlit>=2.8.3",
]

[project.scripts]
patienthub = "patienthub.cli.__main__:app"

[tool.setuptools.packages.find]
include = ["patienthub*"]
exclude = ["data*"]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"
